+++ CHANGELOG.rst
@@ -8,8 +8,9 @@ Since last release
 **Added:**
 
 * Added a warning for when a facility trades with itself (#1895)
+* Added a unit value to Resource, Material, and Product objects (#1894)
 * Added clang-format protections to .cycpp.h and query_backend.h files, modified .clang-format file (#1880)
-* Added a function to facility_cost.cycpp.h to calculate unit price of a DRE bid (#1870 ,#1877, #1884, #1889)
+* Added a function to facility_cost.cycpp.h to calculate unit price of a DRE bid (#1870 ,#1877, #1884, #1889, #1890)
 * Added code injection for matl_buy/sell_policy (#1866)
 * Added set of basic finance math function to EconomicEntity (#1864)
 * Added Composition functions to get printable material composition information (#1868)
@@ -33,7 +34,9 @@ Since last release
 * Allow multiple archetype blocks to facilitate includes (#1874)
 
 **Changed:**
+
 * Changed Dockerfile to use boost and boost-cpp instead of libboost-devel (#1906)
+* Changed TradeExecutor to use adjusted preferences from ExchangeContext (#1897)
 * Ran clang-format on src directory (#1881, #1893)
 * Changed README.rst installation instructions, tested on fresh Ubuntu-22.04 system with Python3.11 (#1744)
 * Rely on ``python3`` in environment instead of ``python`` (#1747)